(PC-22613)[PRO] refactor: remove download-filled icon
