+++ pro/src/icons/index.ts
@@ -45,8 +45,6 @@ export { ReactComponent as DeskIcon } from 'icons/ico-desk.svg'
 export { ReactComponent as DisconnectFullIcon } from 'icons/ico-disconnect-full.svg'
 export { ReactComponent as DisconnectIcon } from 'icons/ico-deconnect.svg'
 export { ReactComponent as DoubleCheckIcon } from 'icons/ico-double-check.svg'
-export { ReactComponent as DownloadFilledIcon } from 'icons/ico-download-filled.svg'
-export { ReactComponent as DownloadIcon } from 'icons/ico-download.svg'
 export { ReactComponent as DropdownDisclosureDownBDisabledIcon } from 'icons/dropdown-disclosure-down-b-disabled.svg'
 export { ReactComponent as DropdownDisclosureDownBIcon } from 'icons/dropdown-disclosure-down-b.svg'
 export { ReactComponent as DropdownDisclosureDownSWIcon } from 'icons/dropdown-disclosure-down-s-w.svg'
